Adds reverse geocoding and changes coords ordering
The coords method returned coordinates in longitude, latitude order.
Since this is what I need for PostGIS this made sense, but not only is
this not the typical format - which I knew quite well - it's not even
the current GIS standard. It's a legacy format. So replacing this with a
sensible default.